Move all remaining specimen actions (i.e., specimen comments) to specimen module. Eliminate old specimen controller.
Clean up SpecimenUrls and SpecimenMigrationService. Move SpecimenUrls impl to specimen module.

Introduce new SpecimenManager to specimen module and move many methods.
