+++ specimen/src/org/labkey/specimen/actions/UpdateSpecimenCommentsBean.java
@@ -4,7 +4,6 @@ import org.apache.commons.lang3.time.DateUtils;
 import org.jetbrains.annotations.NotNull;
 import org.labkey.api.annotations.Migrate;
 import org.labkey.api.data.Container;
-import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.model.SpecimenComment;
 import org.labkey.api.study.Study;
@@ -13,10 +12,10 @@ import org.labkey.api.study.StudyService;
 import org.labkey.api.study.StudyUrls;
 import org.labkey.api.study.TimepointType;
 import org.labkey.api.study.Visit;
-import org.labkey.api.study.model.VisitService;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.RedirectException;
 import org.labkey.api.view.ViewContext;
+import org.labkey.specimen.SpecimenManager;
 
 import java.util.Collections;
 import java.util.Date;
