+++ specimen/src/org/labkey/specimen/report/participant/ParticipantSiteReportFactory.java
@@ -18,17 +18,17 @@ package org.labkey.specimen.report.participant;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.security.User;
-import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
-import org.labkey.specimen.report.SpecimenVisitReport;
-import org.labkey.specimen.report.SpecimenVisitReportAction;
-import org.labkey.specimen.report.SpecimenVisitReportParameters;
-import org.labkey.specimen.actions.SpecimenReportActions;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.Visit;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.Pair;
+import org.labkey.specimen.SpecimenManager;
+import org.labkey.specimen.actions.SpecimenReportActions;
+import org.labkey.specimen.report.SpecimenVisitReport;
+import org.labkey.specimen.report.SpecimenVisitReportAction;
+import org.labkey.specimen.report.SpecimenVisitReportParameters;
 
 import java.util.ArrayList;
 import java.util.Collections;
