+++ specimen/src/org/labkey/specimen/report/participant/ParticipantSummaryReportFactory.java
@@ -16,13 +16,13 @@
 package org.labkey.specimen.report.participant;
 
 import org.labkey.api.data.SimpleFilter;
-import org.labkey.api.specimen.SpecimenManager;
+import org.labkey.api.study.StudyService;
+import org.labkey.api.study.Visit;
+import org.labkey.specimen.SpecimenManager;
+import org.labkey.specimen.actions.SpecimenReportActions;
 import org.labkey.specimen.report.SpecimenVisitReport;
 import org.labkey.specimen.report.SpecimenVisitReportAction;
 import org.labkey.specimen.report.SpecimenVisitReportParameters;
-import org.labkey.specimen.actions.SpecimenReportActions;
-import org.labkey.api.study.StudyService;
-import org.labkey.api.study.Visit;
 
 import java.util.Collections;
 import java.util.List;
