+++ specimen/src/org/labkey/specimen/report/request/RequestParticipantReportFactory.java
@@ -18,10 +18,6 @@ package org.labkey.specimen.report.request;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.security.User;
-import org.labkey.api.specimen.SpecimenManager;
-import org.labkey.specimen.report.SpecimenVisitReport;
-import org.labkey.specimen.report.SpecimenVisitReportAction;
-import org.labkey.specimen.actions.SpecimenReportActions;
 import org.labkey.api.study.Cohort;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
@@ -32,6 +28,10 @@ import org.labkey.api.study.model.VisitService;
 import org.labkey.api.util.DemoMode;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.Pair;
+import org.labkey.specimen.SpecimenManager;
+import org.labkey.specimen.actions.SpecimenReportActions;
+import org.labkey.specimen.report.SpecimenVisitReport;
+import org.labkey.specimen.report.SpecimenVisitReportAction;
 
 import java.util.ArrayList;
 import java.util.Collection;
