+++ specimen/src/org/labkey/specimen/view/requestSpecimens.jsp
@@ -16,7 +16,6 @@
  */
 %>
 <%@ page import="org.labkey.api.data.Container"%>
-<%@ page import="org.labkey.api.specimen.SpecimenManager"%>
 <%@ page import="org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput"%>
 <%@ page import="org.labkey.api.specimen.Vial"%>
 <%@ page import="org.labkey.api.specimen.location.LocationImpl"%>
@@ -28,6 +27,7 @@
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.ViewContext" %>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
+<%@ page import="org.labkey.specimen.SpecimenManager" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.CreateSpecimenRequestForm" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.HandleCreateSpecimenRequestAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.NewRequestBean" %>
