+++ study/api-src/org/labkey/api/specimen/SpecimenManagerNew.java
@@ -12,23 +12,16 @@ import org.labkey.api.data.TableSelector;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.User;
-import org.labkey.api.specimen.model.AdditiveType;
-import org.labkey.api.specimen.model.DerivativeType;
 import org.labkey.api.specimen.model.PrimaryType;
 import org.labkey.api.specimen.model.SpecimenTypeSummary;
 import org.labkey.api.specimen.model.SpecimenTypeSummaryRow;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
-import org.labkey.api.util.DateUtil;
 import org.labkey.api.view.NotFoundException;
 
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.Date;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -52,14 +45,6 @@ public class SpecimenManagerNew
         return !selector.exists();
     }
 
-    public List<Vial> getVials(Container container, User user, String participantId, Double visit)
-    {
-        SimpleFilter filter = SimpleFilter.createContainerFilter(container);
-        filter.addClause(new SimpleFilter.SQLClause("LOWER(ptid) = LOWER(?)", new Object[] {participantId}, FieldKey.fromParts("ptid")));
-        filter.addCondition(FieldKey.fromParts("VisitValue"), visit);
-        return getVials(container, user, filter);
-    }
-
     public List<Vial> getVials(Container container, User user, Set<Long> vialRowIds)
     {
         // Take a set to eliminate dups - issue 26940
@@ -79,36 +64,6 @@ public class SpecimenManagerNew
         return vials;
     }
 
-    public List<Vial> getVials(Container container, User user, String participantId, Date date)
-    {
-        SimpleFilter filter = new SimpleFilter(new SimpleFilter.SQLClause("LOWER(ptid) = LOWER(?)", new Object[] {participantId}, FieldKey.fromParts("ptid")));
-        Calendar endCal = DateUtil.newCalendar(date.getTime());
-        endCal.add(Calendar.DATE, 1);
-        filter.addClause(new SimpleFilter.SQLClause("DrawTimestamp >= ? AND DrawTimestamp < ?", new Object[] {date, endCal.getTime()}));
-        return getVials(container, user, filter);
-    }
-
-    public List<Vial> getVials(Container container, User user, int[] vialsRowIds)
-    {
-        Set<Long> uniqueRowIds = new HashSet<>(vialsRowIds.length);
-        for (int vialRowId : vialsRowIds)
-            uniqueRowIds.add((long)vialRowId);
-        return getVials(container, user, uniqueRowIds);
-    }
-
-    public List<Vial> getVials(Container container, User user, String[] globalUniqueIds) throws SpecimenRequestException
-    {
-        SimpleFilter filter = new SimpleFilter();
-        Set<String> uniqueRowIds = new HashSet<>(globalUniqueIds.length);
-        Collections.addAll(uniqueRowIds, globalUniqueIds);
-        List<String> ids = new ArrayList<>(uniqueRowIds);
-        filter.addInClause(FieldKey.fromParts("GlobalUniqueId"), ids);
-        List<Vial> vials = getVials(container, user, filter);
-        if (vials == null || vials.size() != ids.size())
-            throw new SpecimenRequestException("Vial not found.");       // an id has no matching specimen, let caller determine what to report
-        return vials;
-    }
-
     public List<Vial> getVials(final Container container, final User user, SimpleFilter filter)
     {
         // TODO: LinkedList?
@@ -120,7 +75,7 @@ public class SpecimenManagerNew
         return vials;
     }
 
-    public TableSelector getSpecimensSelector(final Container container, final User user, SimpleFilter filter)
+    private TableSelector getSpecimensSelector(final Container container, final User user, SimpleFilter filter)
     {
         Study study = StudyService.get().getStudy(container);
         if (study == null)
@@ -161,39 +116,6 @@ public class SpecimenManagerNew
         }
         else
             return matches.get(0);
-    }
-
-    @Nullable
-    public AdditiveType getAdditiveType(Container c, int rowId)
-    {
-        List<AdditiveType> additiveTypes = getAdditiveTypes(c, new SimpleFilter(FieldKey.fromParts("RowId"), rowId));
-        if (!additiveTypes.isEmpty())
-            return additiveTypes.get(0);
-        return null;
-    }
-
-    private List<AdditiveType> getAdditiveTypes(final Container container, @Nullable SimpleFilter filter)
-    {
-        final List<AdditiveType> additiveTypes = new ArrayList<>();
-        new TableSelector(SpecimenSchema.get().getTableInfoSpecimenAdditive(container), filter, null).
-                forEachMap(map -> additiveTypes.add(new AdditiveType(container, map)));
-        return additiveTypes;
-    }
-
-    public DerivativeType getDerivativeType(Container c, int rowId)
-    {
-        List<DerivativeType> derivativeTypes = getDerivativeTypes(c, new SimpleFilter(FieldKey.fromParts("RowId"), rowId));
-        if (!derivativeTypes.isEmpty())
-            return derivativeTypes.get(0);
-        return null;
-    }
-
-    private List<DerivativeType> getDerivativeTypes(final Container container, @Nullable SimpleFilter filter)
-    {
-        final List<DerivativeType> derivativeTypes = new ArrayList<>();
-        new TableSelector(SpecimenSchema.get().getTableInfoSpecimenDerivative(container), filter, null).
-                forEachMap(map -> derivativeTypes.add(new DerivativeType(container, map)));
-        return derivativeTypes;
     }
 
     public PrimaryType getPrimaryType(Container c, int rowId)
