+++ study/api-src/org/labkey/api/specimen/SpecimenMigrationService.java
@@ -1,11 +1,10 @@
 package org.labkey.api.specimen;
 
-import org.jetbrains.annotations.Nullable;
 import org.labkey.api.data.Container;
+import org.labkey.api.data.TableInfo;
 import org.labkey.api.services.ServiceRegistry;
 import org.labkey.api.specimen.model.SpecimenRequestEvent;
 import org.labkey.api.view.ActionURL;
-import org.springframework.web.servlet.mvc.Controller;
 
 // Temporary service that provides entry points to ease migration of code from study module to specimen module
 // These should all go away once the migration is complete
@@ -22,17 +21,10 @@ public interface SpecimenMigrationService
     }
 
     ActionURL getBeginURL(Container c);
-    ActionURL getManageRequestStatusURL(Container c, int requestId);
-    ActionURL getManageRequestURL(Container c, int requestId, @Nullable ActionURL returnUrl);
+    ActionURL getInsertSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
     ActionURL getSelectedSpecimensURL(Container c);
     ActionURL getSpecimenEventsURL(Container c, ActionURL returnUrl);
     ActionURL getSpecimenRequestEventDownloadURL(SpecimenRequestEvent event, String name);
     ActionURL getSpecimensURL(Container c);
-    ActionURL getSpecimensURL(Container c, boolean showVials);
-    ActionURL getUploadSpecimensURL(Container c);
-    ActionURL getViewRequestsURL(Container c);
-
-    Class<? extends Controller> getClearCommentsActionClass();
-    Class<? extends Controller> getShowCreateSpecimenRequestActionClass();
-    Class<? extends Controller> getUpdateCommentsActionClass();
+    ActionURL getUpdateSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
 }
