+++ study/src/org/labkey/study/StudyInternalServiceImpl.java
@@ -2,22 +2,34 @@ package org.labkey.study;
 
 import org.labkey.api.data.ActionButton;
 import org.labkey.api.data.Container;
+import org.labkey.api.data.SimpleFilter;
+import org.labkey.api.data.TableInfo;
+import org.labkey.api.query.FieldKey;
+import org.labkey.api.query.QueryUpdateForm;
+import org.labkey.api.query.QueryView;
+import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.SecurityManager.ViewFactory;
 import org.labkey.api.security.User;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.Dataset;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyInternalService;
+import org.labkey.api.study.StudyService;
 import org.labkey.api.study.TimepointType;
 import org.labkey.api.study.Visit;
 import org.labkey.api.study.model.ParticipantDataset;
 import org.labkey.api.study.model.ParticipantInfo;
+import org.labkey.api.view.DataView;
 import org.labkey.api.view.ViewContext;
 import org.labkey.study.model.ParticipantGroupManager;
 import org.labkey.study.model.SecurityType;
 import org.labkey.study.model.StudyImpl;
 import org.labkey.study.model.StudyManager;
 import org.labkey.study.model.VisitImpl;
+import org.labkey.study.query.DatasetQuerySettings;
+import org.labkey.study.query.DatasetQueryView;
+import org.labkey.study.query.SpecimenDetailTable;
+import org.springframework.validation.BindException;
 
 import java.math.BigDecimal;
 import java.util.Collection;
@@ -74,7 +86,7 @@ public class StudyInternalServiceImpl implements StudyInternalService
     @Override
     public String getParticipantCommentProperty(Study study)
     {
-        return null;
+        return ((StudyImpl)study).getParticipantCommentProperty();
     }
 
     @Override
@@ -149,5 +161,41 @@ public class StudyInternalServiceImpl implements StudyInternalService
     public ActionButton createParticipantGroupButton(ViewContext context, String dataRegionName, CohortFilter cohortFilter, boolean hasCreateGroupFromSelection)
     {
         return ParticipantGroupManager.getInstance().createParticipantGroupButton(context, dataRegionName, cohortFilter, hasCreateGroupFromSelection);
+    }
+
+    @Override
+    public QueryView getDatasetQueryView(ViewContext ctx, Study study, User user, Dataset ds, String participantId, int visitId, BindException errors)
+    {
+        UserSchema querySchema = StudyService.get().getStudyQuerySchema(study, user);
+        DatasetQuerySettings qs = (DatasetQuerySettings)querySchema.getSettings(ctx, DatasetQueryView.DATAREGION, ds.getName());
+
+        return new DatasetQueryView(querySchema, qs, errors)
+        {
+            @Override
+            public DataView createDataView()
+            {
+                DataView view = super.createDataView();
+
+                SimpleFilter filter = (SimpleFilter) view.getRenderContext().getBaseFilter();
+                if (null == filter)
+                {
+                    filter = new SimpleFilter();
+                    view.getRenderContext().setBaseFilter(filter);
+                }
+                filter.addCondition(StudyService.get().getSubjectColumnName(view.getViewContext().getContainer()), participantId);
+                if (visitId != 0)
+                    filter.addCondition(FieldKey.fromParts("sequenceNum"), visitId);
+
+                return view;
+            }
+        };
+    }
+
+    @Override
+    public void fixSpecimenRequestableColumn(QueryUpdateForm tableForm)
+    {
+        TableInfo tableInfo = tableForm.getTable(); //TODO: finish fixing bug
+        if (tableInfo instanceof SpecimenDetailTable)
+            ((SpecimenDetailTable)tableInfo).changeRequestableColumn();
     }
 }
