--- study/src/org/labkey/study/controllers/specimen/SpecimenControllerOld.java
@@ -1,313 +1,0 @@
-/*
- * Copyright (c) 2014-2019 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.labkey.study.controllers.specimen;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.labkey.api.action.SimpleViewAction;
-import org.labkey.api.data.BaseColumnInfo;
-import org.labkey.api.data.ButtonBar;
-import org.labkey.api.data.Container;
-import org.labkey.api.data.SimpleFilter;
-import org.labkey.api.data.TableInfo;
-import org.labkey.api.query.FieldKey;
-import org.labkey.api.query.QueryParam;
-import org.labkey.api.query.QueryUpdateForm;
-import org.labkey.api.query.QueryView;
-import org.labkey.api.query.UserSchema;
-import org.labkey.api.query.UserSchemaAction;
-import org.labkey.api.security.RequiresPermission;
-import org.labkey.api.security.permissions.ReadPermission;
-import org.labkey.api.specimen.SpecimenMigrationService;
-import org.labkey.api.specimen.actions.ParticipantCommentForm;
-import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
-import org.labkey.api.study.Dataset;
-import org.labkey.api.study.SpecimenUrls;
-import org.labkey.api.study.Study;
-import org.labkey.api.study.StudyInternalService;
-import org.labkey.api.study.StudyService;
-import org.labkey.api.view.ActionURL;
-import org.labkey.api.view.DataView;
-import org.labkey.api.view.HtmlView;
-import org.labkey.api.view.HttpView;
-import org.labkey.api.view.InsertView;
-import org.labkey.api.view.NavTree;
-import org.labkey.api.view.UpdateView;
-import org.labkey.study.controllers.BaseStudyController;
-import org.labkey.study.query.DatasetQuerySettings;
-import org.labkey.study.query.DatasetQueryView;
-import org.labkey.study.query.SpecimenDetailTable;
-import org.springframework.validation.BindException;
-import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.Controller;
-
-import java.sql.ResultSet;
-
-/**
- * User: brittp
- * Date: Dec 20, 2007
- * Time: 11:08:31 AM
- */
-@SuppressWarnings("UnusedDeclaration")
-public class SpecimenControllerOld extends BaseStudyController
-{
-    private static final Logger _log = LogManager.getLogger(SpecimenControllerOld.class);
-
-    private static final DefaultActionResolver _actionResolver = new DefaultActionResolver(
-        SpecimenControllerOld.class,
-        ParticipantCommentAction.SpecimenCommentInsertAction.class,
-        ParticipantCommentAction.SpecimenCommentUpdateAction.class
-    );
-
-    public SpecimenControllerOld()
-    {
-        setActionResolver(_actionResolver);
-    }
-
-    public static class SpecimenUrlsImpl implements SpecimenUrls
-    {
-        @Override
-        public ActionURL getSpecimensURL(Container c)
-        {
-            return SpecimenMigrationService.get().getSpecimensURL(c);
-        }
-
-        @Override
-        public ActionURL getSpecimensURL(Container c, boolean showVials)
-        {
-            return SpecimenMigrationService.get().getSpecimensURL(c, showVials);
-        }
-
-        @Override
-        public ActionURL getManageRequestURL(Container c, int requestId)
-        {
-            return SpecimenMigrationService.get().getManageRequestURL(c, requestId, null);
-        }
-
-        @Override
-        public ActionURL getRequestDetailsURL(Container c, int requestId)
-        {
-            return getManageRequestURL(c, requestId);
-        }
-
-        @Override
-        public ActionURL getManageRequestStatusURL(Container c, int requestId)
-        {
-            return SpecimenMigrationService.get().getManageRequestStatusURL(c, requestId);
-        }
-
-        @Override
-        public ActionURL getUpdateSpecimenQueryRowURL(Container c, String schemaName, TableInfo table)
-        {
-            ActionURL url = new ActionURL(UpdateSpecimenQueryRowAction.class, c);
-            url.addParameter("schemaName", schemaName);
-            url.addParameter(QueryView.DATAREGIONNAME_DEFAULT + "." + QueryParam.queryName, table.getName());
-
-            return url;
-        }
-
-        @Override
-        public ActionURL getInsertSpecimenQueryRowURL(Container c, String schemaName, TableInfo table)
-        {
-            ActionURL url = new ActionURL(InsertSpecimenQueryRowAction.class, c);
-            url.addParameter("schemaName", schemaName);
-            url.addParameter(QueryView.DATAREGIONNAME_DEFAULT + "." + QueryParam.queryName, table.getName());
-
-            return url;
-        }
-
-        @Override
-        public Class<? extends Controller> getCopyParticipantCommentActionClass()
-        {
-            return CopyParticipantCommentAction.class;
-        }
-    }
-
-    @RequiresPermission(ReadPermission.class)
-    public class CopyParticipantCommentAction extends SimpleViewAction<ParticipantCommentForm>
-    {
-        @Override
-        public ModelAndView getView(final ParticipantCommentForm form, BindException errors)
-        {
-            Study study = getStudyRedirectIfNull();
-            Dataset ds;
-
-            if (form.getVisitId() != 0)
-            {
-                ds = StudyService.get().getDataset(study.getContainer(), StudyInternalService.get().getParticipantVisitCommentDatasetId(study));
-            }
-            else
-            {
-                ds = StudyService.get().getDataset(study.getContainer(), StudyInternalService.get().getParticipantCommentDatasetId(study));
-            }
-
-            if (ds != null)
-            {
-                UserSchema querySchema = StudyService.get().getStudyQuerySchema(study, getUser());
-                DatasetQuerySettings qs = (DatasetQuerySettings)querySchema.getSettings(getViewContext(), DatasetQueryView.DATAREGION, ds.getName());
-                DatasetQueryView queryView = new DatasetQueryView(querySchema, qs, errors)
-                {
-                    @Override
-                    public DataView createDataView()
-                    {
-                        DataView view = super.createDataView();
-
-                        SimpleFilter filter = (SimpleFilter) view.getRenderContext().getBaseFilter();
-                        if (null == filter)
-                        {
-                            filter = new SimpleFilter();
-                            view.getRenderContext().setBaseFilter(filter);
-                        }
-                        filter.addCondition(StudyService.get().getSubjectColumnName(view.getViewContext().getContainer()), form.getParticipantId());
-                        if (form.getVisitId() != 0)
-                            filter.addCondition(FieldKey.fromParts("sequenceNum"), form.getVisitId());
-
-                        return view;
-                    }
-                };
-
-                String lsid = getExistingComment(queryView);
-                ActionURL url;
-                if (lsid != null)
-                    url = new ActionURL(ParticipantCommentAction.SpecimenCommentUpdateAction.class, getContainer()).
-                            addParameter(ParticipantCommentForm.params.lsid, lsid);
-                else
-                    url = new ActionURL(ParticipantCommentAction.SpecimenCommentInsertAction.class, getContainer()).
-                            addParameter(ParticipantCommentForm.params.participantId, form.getParticipantId());
-
-                url.addParameter(ParticipantCommentForm.params.datasetId, ds.getDatasetId()).
-                        addParameter(ParticipantCommentForm.params.comment, form.getComment()).
-                        addReturnURL(form.getReturnActionURL()).
-                        addParameter(ParticipantCommentForm.params.visitId, form.getVisitId());
-
-                for (int rowId : form.getVialCommentsToClear())
-                    url.addParameter(ParticipantCommentForm.params.vialCommentsToClear, rowId);
-                return HttpView.redirect(url);
-            }
-            return new HtmlView("Dataset could not be found");
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-        }
-
-        private String getExistingComment(QueryView queryView)
-        {
-            final TableInfo table = queryView.getTable();
-            if (table != null)
-            {
-                try (ResultSet rs = queryView.getResultSet())
-                {
-                    while (rs.next())
-                    {
-                        String lsid = rs.getString("lsid");
-                        if (lsid != null)
-                            return lsid;
-                    }
-                }
-                catch (Exception e)
-                {
-                    _log.error("Error encountered trying to get " + StudyService.get().getSubjectNounSingular(getContainer()) + " comments", e);
-                }
-            }
-            return null;
-        }
-    }
-
-    @RequiresPermission(EditSpecimenDataPermission.class)
-    public static class UpdateSpecimenQueryRowAction extends UserSchemaAction
-    {
-        @Override
-        public ModelAndView getView(QueryUpdateForm tableForm, boolean reshow, BindException errors)
-        {
-            // Don't allow GlobalUniqueId to be edited
-            TableInfo tableInfo = tableForm.getTable();
-            if (null != tableInfo.getColumn("GlobalUniqueId"))
-                ((BaseColumnInfo)tableInfo.getColumn("GlobalUniqueId")).setReadOnly(true);
-
-            var vialComments = tableInfo.getColumn("VialComments");
-            if (null != vialComments)
-            {
-                ((BaseColumnInfo)vialComments).setUserEditable(false);
-                ((BaseColumnInfo)vialComments).setHidden(true);
-            }
-
-            SpecimenControllerOld.fixSpecimenRequestableColumn(tableForm);
-            ButtonBar bb = createSubmitCancelButtonBar(tableForm);
-            UpdateView view = new UpdateView(tableForm, errors);
-            view.getDataRegion().setButtonBar(bb);
-            return view;
-        }
-
-        @Override
-        public boolean handlePost(QueryUpdateForm tableForm, BindException errors)
-        {
-            doInsertUpdate(tableForm, errors, false);
-            return 0 == errors.getErrorCount();
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-            super.addNavTrail(root);
-            root.addChild("Edit " + _form.getQueryName());
-        }
-    }
-
-    @RequiresPermission(EditSpecimenDataPermission.class)
-    public static class InsertSpecimenQueryRowAction extends UserSchemaAction
-    {
-        @Override
-        public ModelAndView getView(QueryUpdateForm tableForm, boolean reshow, BindException errors)
-        {
-            TableInfo tableInfo = tableForm.getTable();
-            var vialComments = tableInfo.getColumn("VialComments");
-            if (null != vialComments)
-            {
-                ((BaseColumnInfo)vialComments).setUserEditable(false);
-                ((BaseColumnInfo)vialComments).setHidden(true);
-            }
-
-            SpecimenControllerOld.fixSpecimenRequestableColumn(tableForm);
-            InsertView view = new InsertView(tableForm, errors);
-            view.getDataRegion().setButtonBar(createSubmitCancelButtonBar(tableForm));
-            return view;
-        }
-
-        @Override
-        public boolean handlePost(QueryUpdateForm tableForm, BindException errors)
-        {
-            doInsertUpdate(tableForm, errors, true);
-            return 0 == errors.getErrorCount();
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-            super.addNavTrail(root);
-            root.addChild("Insert " + _form.getQueryName());
-        }
-    }
-
-    private static void fixSpecimenRequestableColumn(QueryUpdateForm tableForm)
-    {
-        TableInfo tableInfo = tableForm.getTable(); //TODO: finish fixing bug
-        if (tableInfo instanceof SpecimenDetailTable)
-            ((SpecimenDetailTable)tableInfo).changeRequestableColumn();
-    }
-}
