+++ study/src/org/labkey/study/reports/StudyCrosstabReport.java
@@ -22,6 +22,7 @@ import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.query.QueryParam;
 import org.labkey.api.reports.report.ReportDescriptor;
 import org.labkey.api.reports.report.view.ReportQueryView;
+import org.labkey.api.study.Dataset;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyUtils;
 import org.labkey.api.study.reports.CrosstabReport;
@@ -75,7 +76,7 @@ public class StudyCrosstabReport extends CrosstabReport
     @Override
     public ActionURL getRunReportURL(ViewContext context)
     {
-        int datasetId = NumberUtils.toInt(getDescriptor().getProperty(DatasetDefinition.DATASETKEY), -1);
+        int datasetId = NumberUtils.toInt(getDescriptor().getProperty(Dataset.DATASETKEY), -1);
         if (datasetId == -1)
         {
             String queryName = getDescriptor().getProperty(ReportDescriptor.Prop.queryName);
@@ -94,7 +95,7 @@ public class StudyCrosstabReport extends CrosstabReport
         if (datasetId != -1)
         {
             return new ActionURL(StudyController.DatasetReportAction.class, context.getContainer()).
-                        addParameter(DatasetDefinition.DATASETKEY, datasetId).
+                        addParameter(Dataset.DATASETKEY, datasetId).
                         addParameter(StudyController.DATASET_REPORT_ID_PARAMETER_NAME, getDescriptor().getReportId().toString());
         }
         return super.getRunReportURL(context);
