+++ study/src/org/labkey/study/view/updateQCState.jsp
@@ -19,6 +19,7 @@
 <%@ page import="org.labkey.api.data.DataRegionSelection" %>
 <%@ page import="org.labkey.api.qc.QCState" %>
 <%@ page import="org.labkey.api.qc.QCStateManager" %>
+<%@ page import="org.labkey.api.study.Dataset" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
@@ -27,7 +28,6 @@
 <%@ page import="org.labkey.study.controllers.StudyController.DatasetAction" %>
 <%@ page import="org.labkey.study.controllers.StudyController.UpdateQCStateAction" %>
 <%@ page import="org.labkey.study.controllers.StudyController.UpdateQCStateForm" %>
-<%@ page import="org.labkey.study.model.DatasetDefinition" %>
 <%@ page import="java.util.List" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
@@ -82,7 +82,7 @@
         <tr>
             <td>&nbsp;</td>
             <td><%= button("Update Status").submit(true) %> <%= button("Cancel").href(new ActionURL(DatasetAction.class,
-                    container).addParameter(DatasetDefinition.DATASETKEY, bean.getDatasetId())) %></td>
+                    container).addParameter(Dataset.DATASETKEY, bean.getDatasetId())) %></td>
         </tr>
     </table>
 </labkey:form>
